Работа над [UEE-444] При перепроведении прайс-листа проверять наличие данного прайс-листа в Inbound: сделано.

+++ src/frmUEEMain.cs
@@ -3767,8 +3767,11 @@ and not Exists(select * from farm.blockedprice bp where bp.PriceCode = ?DeletePr
 					{
 						if (File.Exists(rootpath + "Base\\" + CurrentFileName))
 						{
-							File.Copy(rootpath + "Base\\" + CurrentFileName, rootpath + "Inbound0\\" + CurrentFileName);
-							PricesRetrans(now, RetransedPriceList[CurrentPriceCode].PriceCode);
+							if (!File.Exists(rootpath + "Inbound0\\" + CurrentFileName))
+							{
+								File.Copy(rootpath + "Base\\" + CurrentFileName, rootpath + "Inbound0\\" + CurrentFileName);
+								PricesRetrans(now, RetransedPriceList[CurrentPriceCode].PriceCode);
+							}
 						}
 						RetransedPriceList.RemoveAt(CurrentPriceCode);
 					}
